+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - Unreleased
+
+[0.4.0]: https://github.com/HMIProject/open62541/compare/v0.3.0...v0.4.0
+
+### Added
+
+- Fallible conversion from `time::OffsetDateTime` to `ua::DateTime`.
+
+### Changed
+
+- Breaking: Renamed `ua::DateTime::as_datetime()` to `ua::DateTime::to_utc()`.
+
 ## [0.3.0] - 2024-01-23
 
 [0.3.0]: https://github.com/HMIProject/open62541/compare/v0.2.2...v0.3.0